review: transformable camera
- add/improve documentation
- move helper function _resolution_after_transform to USBCamera (it's only used there!)
